拆分返回值实现方法
