+++ src/TelescopeSourceGenerator/Analyzers/ExportTypeToMethod_/IExportMethodReturnTypeInfo.cs
@@ -1,5 +1,10 @@
 ﻿namespace dotnetCampus.Telescope.SourceGeneratorAnalyzers;
 
+/// <summary>
+/// 导出方法返回值类型信息
+/// </summary>
+/// 比如属于枚举的 ValueTuple 返回值类型的
+/// 比如属于数组返回值类型的
 public interface IExportMethodReturnTypeInfo
 {
 }
